Rename Ethereum LA file
