+++ packages/wrapped-keys/src/index.ts
@@ -1,5 +1,5 @@
 export * from './lib/wrapped-keys';
-export * from './lib/litActions/litActions';
+export * from './lib/litActions/ethereum';
 export * from './lib/interfaces';
 
 export const signTransactionWithSolanaEncryptedKeyLitActionIpfsCid =
